Rename App to LoggedInApp
