+++ src/entries/components/LoggedInApp/LoggedInApp.spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import { App } from './App';
+import { App } from './LoggedInApp';
 import Header from './header';
 import Footer from './footer';
