+++ src/entries/components/LoggedInApp/footer/Footer.js
