+++ src/entries/components/LoggedInApp/footer/Footer.spec.js
