+++ src/entries/components/LoggedInApp/footer/index.js
